+++ src/components/LanguageSwitcher.tsx
@@ -1,8 +1,14 @@
 import React from 'react';
 import styled from 'styled-components';
 import { graphql } from 'gatsby';
+import * as i18Next from 'gatsby-plugin-react-i18next';
 
-import { useLocale, Link } from '~/hooks/useLocale';
+type useLocaleHook = {
+  getTranslationFor: (key: string) => string;
+  getCurrentLanguage: () => string;
+  getAllLanguages: () => string[];
+  getOriginalPath: () => string;
+};
 
 const Container = styled.div`
   display: flex;
@@ -29,6 +35,50 @@ export const query = graphql`
     }
   }
 `;
+
+export function useLocale(): useLocaleHook {
+  const { t } = i18Next.useTranslation();
+  const context = i18Next.useI18next();
+
+  function getTranslationFor(key: string) {
+    return t(key);
+  }
+
+  function getCurrentLanguage() {
+    return context.language;
+  }
+
+  function getAllLanguages() {
+    return context.languages;
+  }
+
+  function getOriginalPath() {
+    return context.originalPath;
+  }
+
+  return {
+    getTranslationFor,
+    getCurrentLanguage,
+    getAllLanguages,
+    getOriginalPath,
+  };
+}
+
+export function getContentLanguage(locale: useLocaleHook, langkey?: string | null) {
+  if (!langkey) {
+    return;
+  }
+
+  const languagePrefix = locale.getTranslationFor('article.languagePrefix');
+  const languageName = locale.getTranslationFor(`languages.${langkey}`);
+
+  return `${languagePrefix} ${languageName}`;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function Link(props: any) {
+  return <i18Next.Link {...props} />;
+}
 
 export default function LanguageSwitcher() {
   const locale = useLocale();
