--- src/hooks/__tests__/useTheme.test.js
@@ -1,23 +1,0 @@
-import { renderHook, act } from '@testing-library/react';
-
-import { useTheme } from '~/hooks/useTheme';
-
-describe('useTheme', () => {
-  it('returns the correct properties', async () => {
-    const { result } = renderHook(() => useTheme());
-
-    expect(result.current[0]).toBe('light');
-
-    act(() => {
-      result.current[1]();
-    });
-
-    expect(result.current[0]).toBe('dark');
-
-    act(() => {
-      result.current[1]();
-    });
-
-    expect(result.current[0]).toBe('light');
-  });
-});
