+++ src/pages/__tests__/links.test.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+
+import { useStaticQuery, mockUseStaticQuery } from '../../__mocks__/useStaticQuery';
+import Links from '../links';
+
+jest.mock('../../components/Footer', () => () => <p>Footer</p>);
+
+describe('<Links>', () => {
+  beforeEach(() => {
+    useStaticQuery.mockImplementation(() => mockUseStaticQuery);
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  it('renders properly', () => {
+    const { baseElement } = render(<Links />);
+    expect(baseElement).toMatchSnapshot();
+  });
+});
